// angular-cli/let/src/app/models/question.ts
export class Question {
   id: string;
   guid: string;
   name: string;
   pts: string;
   description: string;
   illustration: string;
   category_majoring_id: string;
   category_majoring: string;
   iscorrect: string;
   my_answer: string;
   answer_name: string;
   my_answer_txt: string;
   idx: string;
}